pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hackmenot"
version = "0.1.0"
description = "AI-Era Code Security Scanner"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "hackmenot team" }
]
keywords = ["security", "scanner", "ai", "code-analysis", "sast"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "tree-sitter>=0.23.0",
    "tree-sitter-javascript>=0.23.0",
    "tree-sitter-go>=0.21.0",
    "python-hcl2>=4.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
hackmenot = "hackmenot.cli.main:app"

[project.urls]
Homepage = "https://github.com/hackmenot/hackmenot"
Repository = "https://github.com/hackmenot/hackmenot"

[tool.hatch.build.targets.wheel]
packages = ["src/hackmenot"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build]
include = [
    "src/hackmenot/**/*.py",
    "src/hackmenot/**/*.yml",
    "src/hackmenot/**/*.txt",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
